% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bb.R
\name{read_bb}
\alias{read_bb}
\title{Parse user data from BlackBoard}
\usage{
read_bb(text = readLines("clipboard"), role = "student")
}
\arguments{
\item{text}{Character string containing BlackBoard user data,
Default: readLines("clipboard")}

\item{role}{Character vector containing roles to retain, Default: 'student'}
}
\value{
A \code{data.frame}
}
\description{
Copy the contents of the BlackBoard users page to the clipboard,
and run this function.
}
\examples{
tmp <- c("\t 48483 Options Menu: Username\tFirst\tLast\tfirstlast@students.uu.nl\tstudent\t\tYes")
read_bb(tmp)
}
